training-api/requirements.txt
apscheduler
databases
fastapi
pyhumps
uvicorn
pydantic
redis
requests
sqlalchemy
sqlalchemy-utils
